third_party/WebKit/LayoutTests/fast/dom/shadow/drag-and-drop-in-user-agent-shadow-expected.txt

Tests whether we can not start dragging a node in user agent shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dragstartCount is 1
PASS successfullyParsed is true

TEST COMPLETE
